requirements.txt
PyQt6>=6.4.0
pyserial>=3.5
pyqtgraph>=0.13.1
matplotlib>=3.7.1
cryptography>=39.0.0
python-dotenv>=0.21.1
QDarkStyle>=3.1
numpy>=1.24.0
pandas>=1.5.3
pymodbus>=3.0.0
setuptools>=69.0.0
pyinstaller>=6.3.0
psutil>=5.9.0
PySide6>=6.4.0 